refactor: change components paths
